requirements.txt
fire
beautifulsoup4
requests
rich
attrs
cattrs
transformers
transformers[torch]
bfloat16
huggingface_hub
streamlit
openai
googlesearch-python
setuptools
